requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.4
annotated-types==0.6.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.40.0
boto3==1.34.78
botocore==1.34.78
branca==0.7.2
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.18.0
debugpy==1.8.1
decorator==4.4.2
diff-match-patch==20230430
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
Django==5.0.4
django-import-export==4.0.7
django-widget-tweaks==1.5.0
easyocr==1.7.1
einops==0.6.1
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
ExifRead==3.0.0
filelock==3.13.1
filterpy==1.4.5
fitz==0.0.1.dev2
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.48.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
ghp-import==2.1.0
google-cloud==0.34.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.2.1
kiwisolver==1.4.5
langchain==0.0.233
langsmith==0.0.5
lazy_loader==0.3
libclang==18.1.1
llama_cpp_python==0.2.44
looseversion==1.3.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.5.12
mkdocs-material-extensions==1.3.1
mkdocs-table-reader-plugin==2.1.0
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.0
ninja==1.11.1.1
nipype==1.8.6
numexpr==2.9.0
numpy==1.26.4
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optimum==1.17.1
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
paginate==0.5.6
pandas==2.2.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
peft==0.9.0
piexif==1.1.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.4.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydot==2.0.0
Pygments==2.17.2
pymdown-extensions==10.7
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.26.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==25.1.2
rdflib==7.0.0
regex==2023.12.25
reportlab==4.2.0
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.18.1
sentencepiece==0.2.0
setuptools==69.1.0
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
tablib==3.5.0
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.2.12
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
traits==6.3.2
transformers==4.37.2
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.43
urllib3==2.2.1
virtualenv==20.25.3
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
whitenoise==6.4.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
